backend/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "commoditywatch-contango-backend"
version = "0.1.0"
description = "Contango backend for CommodityWatch InventoryWatch MVP."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14,<2",
  "apscheduler>=3.10,<4",
  "argon2-cffi>=23.1,<26",
  "asyncpg>=0.30,<1",
  "email-validator>=2.2,<3",
  "fastapi>=0.115,<1",
  "greenlet>=3,<4",
  "httpx>=0.28,<1",
  "orjson>=3.10,<4",
  "psycopg[binary]>=3.2,<4",
  "pydantic>=2.10,<3",
  "pydantic-settings>=2.7,<3",
  "python-dateutil>=2.9,<3",
  "PyYAML>=6,<7",
  "requests>=2.32,<3",
  "beautifulsoup4>=4.12,<5",
  "xlrd>=2.0,<3",
  "sqlalchemy>=2.0.38,<3",
  "uvicorn[standard]>=0.34,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9",
  "pytest-asyncio>=0.25,<1",
  "respx>=0.22,<1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
markers = [
  "asyncio: marks async tests that require pytest-asyncio",
]
